+++ src/paragraph/story.tsx
@@ -9,8 +9,8 @@ const stories = storiesOf('Widgets|Paragraph', module)
 stories.addDecorator(withKnobs)
 
 const shortText = 'Short text (below max char)'
-const longText =
-  'Long text (above max char) - Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
+const longText = 'Long text (above max char) '.repeat(20)
+const longTextWithoutSpaces = `http://${'long_url_without_spaces'.repeat(20)}.com`
 
 stories.add('Long text', () => (
   <Section>
@@ -26,5 +26,11 @@ stories.add('Long text', () => (
 stories.add('Short text', () => (
   <Section>
     <Paragraph>{shortText}</Paragraph>
+  </Section>
+))
+
+stories.add('Long unbreakable text', () => (
+  <Section>
+    <Paragraph>{longTextWithoutSpaces}</Paragraph>
   </Section>
 ))